Send SIGKILL when `offline-imap' is called with prefix
I'm using offlineimap and offlineimap-el to read my mails with
wanderlust in emacs. When offlineimap crashes (mostly due to a lost
internet connection as I am working on a laptop) it crashes hard,
which means I can only kill it with sigkill. Your offlineimap-kill
function only sends sigterm, which is not enough. Today I was annoyed
enough to finally fix it. Now, when I give a prefix arg sigkill is
used. I thought you might want to add that to your git repository
(diff attached).

Signed-off-by: Julien Danjou <julien@danjou.info>